requirements.txt
keyboard==0.13.5
Pillow==9.0.1
PyAutoGUI==0.9.52
